chore(focusmanager): rename focuslock to focusmanager
