+++ packages/dialog/package.json
@@ -28,7 +28,7 @@
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^4.3.5",
     "@pluralsight/ps-design-system-theme": "^1.3.4",
-    "@pluralsight/ps-design-system-focuslock": "^1.0.0",
+    "@pluralsight/ps-design-system-focusmanager": "^1.0.0",
     "polished": "^1.9.2",
     "prop-types": "^15.5.10"
   },
