+++ packages/focusmanager/.npmignore
