+++ packages/focusmanager/.npmrc
