+++ packages/focusmanager/index.js
