+++ packages/focusmanager/src/react/__stories__/index.story.js
@@ -3,10 +3,10 @@ import React from 'react'
 
 import Text from '@pluralsight/ps-design-system-text/react'
 
-import FocusLock from '..'
+import FocusManager from '..'
 
-storiesOf('FocusLock', module).add('default', _ => (
-  <FocusLock style={{ color: 'white', textAlign: 'center', margin: 100 }}>
+storiesOf('FocusManager', module).add('default', _ => (
+  <FocusManager style={{ color: 'white', textAlign: 'center', margin: 100 }}>
     <Text.P>
       <a href="#">Tri-tip pork chop jowl</a>, spare ribs shankle sirloin kevin.
       Ham hock tri-tip jowl, <a href="#">shank venison ham</a> andouille rump
@@ -16,5 +16,5 @@ storiesOf('FocusLock', module).add('default', _ => (
       rump buffalo shoulder chicken beef brisket. Spare ribs drumstick tail
       cupim bacon filet mignon.
     </Text.P>
-  </FocusLock>
+  </FocusManager>
 ))
