+++ packages/focusmanager/vars.js
